+++ Source/Kubernetes/src/main/java/io/dolittle/moose/kubernetes/informers/ICanProvideInformers.java
@@ -1,7 +1,7 @@
 // Copyright (c) Dolittle. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
-package io.dolittle.moose.kubernetes.service;
+package io.dolittle.moose.kubernetes.informers;
 
 import io.kubernetes.client.informer.SharedIndexInformer;
 import io.kubernetes.client.informer.SharedInformerFactory;
