+++ Source/Kubernetes/src/main/java/io/dolittle/moose/kubernetes/informers/SharedIndexInformers.java
@@ -1,7 +1,7 @@
 // Copyright (c) Dolittle. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
-package io.dolittle.moose.kubernetes.service;
+package io.dolittle.moose.kubernetes.informers;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
